+++ packages/charts/src/components/BarChart/BarChart.tsx
@@ -89,11 +89,11 @@ export interface BarChartProps extends IChartBaseProps {
   /**
    * An array of config objects. Each object will define one dimension of the chart.
    *
-   * #### Required Properties
+   * **Required Properties**
    * - `accessor`: string containing the path to the dataset key the dimension should display. Supports object structures by using <code>'parent.child'</code>.
    *   Can also be a getter.
    *
-   * #### Optional Properties
+   * **Optional Properties**
    * - `formatter`: function will be called for each data label and allows you to format it according to your needs
    * - `interval`: number that controls how many ticks are rendered on the x axis
    *
@@ -102,11 +102,11 @@ export interface BarChartProps extends IChartBaseProps {
   /**
    * An array of config objects. Each object is defining one bar in the chart.
    *
-   * #### Required properties
+   * **Required properties**
    * - `accessor`: string containing the path to the dataset key this bar should display. Supports object structures by using <code>'parent.child'</code>.
    *   Can also be a getter.
    *
-   * #### Optional properties
+   * **Optional properties**
    *
    * - `label`: Label to display in legends or tooltips. Falls back to the <code>accessor</code> if not present.
    * - `color`: any valid CSS Color or CSS Variable. Defaults to the `sapChart_OrderedColor_` colors